requirements.txt
flask==2.2.3
flask-cors==3.1.1
google-generativeai==0.1.5
python-dotenv==1.0.0
youtube-transcript-api==0.6.0
